+++ Sources/event/Extensions/EKReminder+LocationAlarms.swift
@@ -1,6 +1,7 @@
 #if canImport(EventKit)
 
   import EventKit
+  import EventModels
 
   // MARK: - Reminder Location Alarms
 
@@ -14,6 +15,15 @@
       for alarm in locationAlarms {
         removeAlarm(alarm)
       }
+    }
+
+    /// Replace the repeat rule. EventKit will not save one on a reminder without a due
+    /// date, so that is checked here for a readable error.
+    func setRecurrenceRule(_ rule: RecurrenceRule) throws {
+      guard dueDateComponents != nil else {
+        throw EventCLIError.invalidInput("A repeat rule needs a due date; set --due as well.")
+      }
+      recurrenceRules = [try rule.toEKRecurrenceRule()]
     }
   }